+++ View/src/main/java/convenience/store/StockReserveCancelled.java
@@ -3,7 +3,7 @@ package convenience.store;
 public class StockReserveCancelled extends AbstractEvent {
 
     private Long id;
-    private Integer reserveId;
+    private Long reserveId;
     private Integer goodId;
     private String goodName;
     private Integer reserveQty;
@@ -16,11 +16,11 @@ public class StockReserveCancelled extends AbstractEvent {
     public void setId(Long id) {
         this.id = id;
     }
-    public Integer getReserveId() {
+    public Long getReserveId() {
         return reserveId;
     }
 
-    public void setReserveId(Integer reserveId) {
+    public void setReserveId(Long reserveId) {
         this.reserveId = reserveId;
     }
     public Integer getGoodId() {
